+++ R/importInput.R
@@ -421,3 +421,63 @@
 #   out <- gsub(" ", "%20", out)
 # }
 
+
+# "st-storage"
+.importSTStorage <- function(area, timeStep, opts, ...){
+
+  if (!area %in% opts$areasWithSTClusters) 
+    return(NULL)
+  
+  if(!"api" %in% opts$typeLoad){
+    clusters <- list.files(file.path(opts$inputPath, "st-storage/series", area))
+  } else {
+    clusters <- names(read_secure_json(file.path(opts$inputPath, "st-storage/series", area), 
+                                       token = opts$token, timeout = opts$timeout, config = opts$httr_config))
+  }
+  
+  # "st-storage" have 5 txt files output for each cluster
+  list_names_txt_files <- list.files(file.path(opts$inputPath, 
+                                               "st-storage/series", area, clusters))
+  list_names_less_txt <- sub(pattern = ".txt", replacement = "", x = list_names_txt_files)
+
+  ldply(clusters, function(cl) {
+    pattern <- paste0("%s/%s/%%s/", list_names_txt_files)
+    filePatterns <- sprintf(pattern, "st-storage/series", area)
+    
+    res <- lapply(filePatterns, 
+                  function(.x){
+                    index_name_file <- which(filePatterns %in% .x)
+                    
+                    res_temp <- .importInputTS(area= cl, 
+                                               timeStep= timeStep, 
+                                               opts= opts, 
+                                               fileNamePattern= .x,
+                                               colnames= "st-storage",
+                                               inputTimeStep = "hourly", 
+                                               type = "matrix")
+                    res_temp$name_file <- list_names_less_txt[index_name_file]
+                    res_temp
+                    })
+    
+    res <- rbindlist(res)
+    
+    # area, timeStep, opts, fileNamePattern, colnames, 
+    # inputTimeStep, fun = "sum", type = "simple", colSelect = NULL, ...
+    
+    # res <- .importInputTS(cl, timeStep, opts, filePattern, "st-storage",
+    #                       inputTimeStep = "hourly", type = "matrix")
+    
+    
+    
+    if (is.null(res)) return(NULL)
+    
+    res$area <- area
+    res$cluster <- cl
+    
+    setcolorder(res, c("area", "cluster", "timeId", setdiff(names(res), 
+                                                            c("area", "cluster", "timeId"))))
+  })
+  
+  # added a column "name_file" to tag the file name
+  
+}